+++ apps/web/src/components/math-skills/number-input-exercise/number-input-exercise.tsx
