+++ apps/web/src/components/math-skills/number-input-exercise/number-keyboard.tsx
