+++ apps/web/src/components/math-skills/number-line-exercise/arrow-button-navigation.tsx
