+++ apps/web/src/components/math-skills/number-line-exercise/number-labels.tsx
