+++ apps/web/src/pages/meine-mathe-skills/[grade]/[exercise].tsx
@@ -1,9 +1,9 @@
 import { NextPage } from 'next'
 import { useRouter } from 'next/router'
 
-import { MathSkillsHeader } from '@/components/content/exercises/math-skills-header'
-import { NumberInputExercise } from '@/components/content/exercises/number-input-exercise/number-input-exercise'
-import { NumberLineExercise } from '@/components/content/exercises/number-line-exercise/number-line-exercise'
+import { MathSkillsHeader } from '@/components/math-skills/math-skills-header'
+import { NumberInputExercise } from '@/components/math-skills/number-input-exercise/number-input-exercise'
+import { NumberLineExercise } from '@/components/math-skills/number-line-exercise/number-line-exercise'
 import { FrontendClientBase } from '@/components/frontend-client-base'
 import { HeadTags } from '@/components/head-tags'
