+++ apps/web/src/pages/meine-mathe-skills/[grade]/index.tsx
@@ -1,7 +1,7 @@
 import { NextPage } from 'next'
 import { useRouter } from 'next/router'
 
-import { MathSkillsHeader } from '@/components/content/exercises/math-skills-header'
+import { MathSkillsHeader } from '@/components/math-skills/math-skills-header'
 import { Link } from '@/components/content/link'
 import { FrontendClientBase } from '@/components/frontend-client-base'
 import { HeadTags } from '@/components/head-tags'
